requirements.txt
lazy-object-proxy==1.4.3
psycopg2==2.8.6
pycryptodome==3.9.9
pycryptodomex==3.9.9
pyDH==0.1.6
socks==0
PySocks==1.7.1
python-socketio==5.0.4